requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# RAG
lightrag-hku==1.0.0

# Vector DB
pymilvus==2.5.4
milvus-lite==2.4.12

# Retry
tenacity==9.0.0

# Graph DB
neo4j==5.27.0

# Queue
redis==5.2.1
rq==2.1.0

# LLM / Embeddings
openai==1.58.1
tiktoken==0.8.0

# Document processing
python-docx==1.1.2
pypdf==5.1.0
beautifulsoup4==4.12.3

# Text processing
langchain-text-splitters==0.3.4

# Config & utils
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Dev & testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
ruff==0.8.6
